+++ Login/Stage2.cs
@@ -23,20 +23,27 @@ namespace EliteDangerousDataProvider
                 Credentials.ToFile();
 
                 CompanionAppService app = new CompanionAppService(Credentials);
-                Commander Cmdr = app.Profile();
+                try
+                {
+                    Commander Cmdr = app.Profile();
+                    if (Cmdr == null)
+                    {
+                        MessageBox.Show("There was a problem.  You will need to close and restart this app and attempt to log in again", "Problem detected", MessageBoxButtons.OK, MessageBoxIcon.Error);
+                        Credentials.Clear();
+                        Credentials.ToFile();
+                    }
+                    else
+                    {
+                        Stage3 stage3 = new Stage3(Cmdr.Name);
+                        stage3.Show();
+                        this.Hide();
+                    }
+                }
+                catch (Exception ex)
+                {
+                    MessageBox.Show("There was a problem with your profile.  Please report the below error to github.com/cmdrmcdonald/EliteDangerousDataProvider/issues\r\n" + ex.StackTrace);
+                }
 
-                if (Cmdr == null)
-                {
-                    MessageBox.Show("There was a problem.  You will need to close and restart this app and attempt to log in again", "Problem detected", MessageBoxButtons.OK, MessageBoxIcon.Error);
-                    Credentials.Clear();
-                    Credentials.ToFile();
-                }
-                else
-                {
-                    Stage3 stage3 = new Stage3(Cmdr.Name);
-                    stage3.Show();
-                    this.Hide();
-                }
             }
         }
     }